feat: add devices mock support
