+++ src/modules/interface/themes/module.ts
@@ -3,7 +3,6 @@ import BlankLayout from "@/layouts/BlankLayout.vue";
 import ThemeLayout from "./layouts/ThemeLayout.vue";
 import ThemeDetail from "./ThemeDetail.vue";
 import ThemeSetting from "./ThemeSetting.vue";
-import Visual from "./Visual.vue";
 import { IconPalette } from "@halo-dev/components";
 import { markRaw } from "vue";
 
@@ -38,20 +37,6 @@ export default definePlugin({
           meta: {
             title: "主题设置",
             permissions: ["system:settings:view"],
-          },
-        },
-      ],
-    },
-    {
-      path: "/theme/visual",
-      component: BlankLayout,
-      children: [
-        {
-          path: "",
-          name: "ThemeVisual",
-          component: Visual,
-          meta: {
-            title: "可视化编辑",
           },
         },
       ],
